fix(control-box): one ref matcher for adopt and pull
hub-pull still exact-matched after adopt learned prefixes, so a ref adopt
accepted could make pull miss the registration, lose `provider`, and write the
keys under the raw ref instead of the box's sandbox id — one box's keys in
another id's dir and custody subtree.

Both now share `match-ref.ts`. It gets its own module precisely because the two
callers key different things off the result, which is how they drifted apart.
