+++ apps/cli/src/control-plane/hub-pull.ts
@@ -13,6 +13,7 @@ import { chmod, mkdir, writeFile } from 'node:fs/promises';
 import { boxSshDirForProvider, defaultBoxSshDir } from '@agentbox/sandbox-core';
 import type { CustodyClient } from './custody-client.js';
 import type { ControlPlaneAdminClient } from './admin-client.js';
+import { matchRegistration } from './match-ref.js';
 
 export interface HubPullResult {
   /** The id the keys are stored under (sandboxId, or the box id as a fallback). */
@@ -40,14 +41,11 @@ export interface HubPullArgs {
  */
 export async function pullBoxSshKeys(args: HubPullArgs): Promise<HubPullResult> {
   const boxes = await args.admin.listBoxes();
-  // Match the same refs adoption does — id, name, AND sandbox id. Matching only
-  // id/name silently lost `provider` for a sandbox-id ref, which sent the keys
-  // to the un-namespaced default dir while the box's identityFile pointed at the
-  // provider-namespaced one (DigitalOcean); hetzner only survived because its
-  // dir is un-namespaced anyway.
-  const reg = boxes.find(
-    (b) => b.boxId === args.box || b.name === args.box || b.sandboxId === args.box,
-  );
+  // The SAME matcher adoption uses. Resolving refs differently here is not a
+  // cosmetic inconsistency: a ref this missed but adoption matched lost
+  // `provider` and fell back to `args.box` as the key — writing one box's keys
+  // into another id's on-disk dir and custody subtree.
+  const reg = matchRegistration(boxes, args.box);
   const key = reg?.sandboxId ?? reg?.boxId ?? args.box;
   const files = await downloadBoxSshKeys({
     custody: args.custody,
